+++ src/examples/radio_arq_leds.c
@@ -9,8 +9,8 @@
 #include "radio_arq.h"
 #include "utils.h"
 
-static uint8_t led_gpios[] = {LED_PIN_2, LED_PIN_4, LED_PIN_3, LED_PIN_1};
-static size_t led_gpios_size = sizeof(led_gpios) / sizeof(uint8_t);
+static uint16_t led_gpios[] = {LED_PIN_2, LED_PIN_4, LED_PIN_3, LED_PIN_1};
+static size_t led_gpios_size = sizeof(led_gpios) / sizeof(uint16_t);
 
 static uint32_t report;
